+++ flink-runtime/src/main/java/org/apache/flink/runtime/scheduler/SchedulerBase.java
@@ -205,8 +205,7 @@ public abstract class SchedulerBase implements SchedulerNG, CheckpointScheduling
                 new ExecutionGraphHandler(executionGraph, log, ioExecutor, this.mainThreadExecutor);
 
         this.operatorCoordinatorHandler =
-                new OperatorCoordinatorHandlerImplementation(
-                        executionGraph, this::handleGlobalFailure);
+                new DefaultOperatorCoordinatorHandler(executionGraph, this::handleGlobalFailure);
         operatorCoordinatorHandler.initializeOperatorCoordinators(this.mainThreadExecutor);
         exceptionHistory =
                 new BoundedFIFOQueue<>(
